Day 4 solution
Includes a start to day 3 solution, a completion of day 4, and some idiomatic syntax changes to day 2.

Day 3 is proving difficult. One way to approach it, and currently the way I'm trying, is to iterate until a special character is found. From the special character, search in a 'box' around the special character.

The way I'm trying to implement this is, for example using the 'search_above' func: special char is found -> return the chars in the line above from (special char index - 1, spec char index, spec char index + 1)

If there is a number in the list of chars found, traverse until a numeric char is not found. This should give the number. This doesn't currently work because the returned slice does not contain all of the possible numbers.

Returning a larger slice that contains all possible numeric indexes relative to the special char gives obvious parsing issues.