Day 5 solution
The meat and potatoes of this solution is in the "solve_given_seed()" function (the rest is just parsing). My first thought after reading the problem would be to iterate and generate all mappings/numbers in the given range.

This quickly proved that it would not be a good solution after looking at the actual puzzle.

Enter... math.

No, it was really simple and should be trivial after looking at the function.

There were also a few refactors on old solutions thanks to cargo clippy.